Blind fix crash when bss is empty.
Try to fix the issue reported in #5. I couldn't repro it right now.

I think the issue is that bss is empty and the state section is well aligned. In this condition realloc will free the data memory as the size will be zero.

Also add some more checks on the copying of section data.